feat: Extract Loading and Error progress indicators to components. Extract registration from LoginViewModel. AuthenticationViewModel -> LoginViewModel.
