+++ app/src/main/java/no/usn/mob3000/domain/viewmodel/auth/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package no.usn.mob3000.domain.viewmodel
+package no.usn.mob3000.domain.viewmodel.auth
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
@@ -12,19 +12,17 @@ import no.usn.mob3000.domain.model.AuthError
 import no.usn.mob3000.domain.model.User
 import no.usn.mob3000.domain.usecase.auth.LoginUseCase
 import no.usn.mob3000.domain.usecase.auth.LogoutUseCase
-import no.usn.mob3000.domain.usecase.auth.RegisterUseCase
 
 /**
  * ViewModel to user state and, if necessary, authentication state.
  *
  * @property loginUseCase The Android Use Case handling login business logic.
- * @author frigvid, Anarox
+ * @author Anarox
  * @created 2024-10-21
  */
 class LoginViewModel(
     private val loginUseCase: LoginUseCase = LoginUseCase(),
-    private val logoutUseCase: LogoutUseCase = LogoutUseCase(),
-    private val registerUseCase: RegisterUseCase = RegisterUseCase()
+    private val logoutUseCase: LogoutUseCase = LogoutUseCase()
 ) : ViewModel() {
     /**
      * The current [LoginState].
@@ -85,6 +83,7 @@ class LoginViewModel(
         _loginState.value = LoginState.Idle
     }
 
+    // TODO: Yeetus deletus.
     fun logout() {
         viewModelScope.launch {
             _authenticatedUser.value = null
@@ -92,38 +91,7 @@ class LoginViewModel(
             logoutUseCase()
         }
     }
-
-    /**
-     * Initiates the registration process with the provided credentials.
-     *
-     * @param email The user's e-mail address.
-     * @param password The user's password.
-     * @author Anarox
-     * @created 2024-10-30
-     */
-
-    fun register(email: String, password: String) {
-        viewModelScope.launch {
-            _authenticatedUser.value = null
-            _loginState.value = LoginState.Loading
-            registerUseCase(email, password).fold(
-                onSuccess = { user ->
-                    _authenticatedUser.value = user
-                    _loginState.value = LoginState.Success(user)
-                },
-                onFailure = { error ->
-                    val authError = when (error) {
-                        is AuthRestException -> AuthError.fromException(error)
-                        else -> AuthError.Unknown(error.message ?: "Unknown error occurred")
-                    }
-
-                    _loginState.value = LoginState.Error(authError)
-                }
-            )
-        }
-    }
 }
-
 
 /**
  * Sealed class representing the possible states of the login process.
