+++ app/src/main/java/no/usn/mob3000/ui/components/auth/Error.kt
@@ -1,0 +1,36 @@
+package no.usn.mob3000.ui.components.auth
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun Error(
+    cardClickable: () -> Unit = {  },
+    text: String,
+    color: Color = MaterialTheme.colorScheme.onErrorContainer
+) {
+    Card(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(vertical = 8.dp)
+            .clickable { cardClickable() },
+        colors = CardDefaults.cardColors(
+            containerColor = MaterialTheme.colorScheme.errorContainer
+        )
+    ) {
+        Text(
+            text = text,
+            color = color,
+            modifier = Modifier.padding(16.dp)
+        )
+    }
+}
